backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sota-statworks"
version = "0.1.0"
description = "AI-powered statistical decision engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "python-pptx>=0.6",
    "openai>=1.12",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]
